formatting and casting

+++ src/arkode/arkode_lsrkstep.c
@@ -424,7 +424,8 @@ int lsrkStep_FullRHS(ARKodeMem ark_mem, sunrealtype t, N_Vector y, N_Vector f,
       If the stepper is an SSP, fn is updated and reused at the beginning 
       of the step unless ark_mem->fn_is_current is changed by ARKODE.
       */
-    if(step_mem->is_SSP){
+    if (step_mem->is_SSP)
+    {
       retval = step_mem->fe(t, y, ark_mem->fn, ark_mem->user_data);
       step_mem->nfe++;
       ark_mem->fn_is_current = SUNTRUE;
@@ -543,7 +544,7 @@ int lsrkStep_TakeStepRKC(ARKodeMem ark_mem, sunrealtype* dsmPtr, int* nflagPtr)
     }
   }
 
-  step_mem->req_stages = SUNRround(ss);
+  step_mem->req_stages = (int)SUNRround(ss);
   step_mem->stage_max  = SUNMAX(step_mem->req_stages, step_mem->stage_max);
 
 #if SUNDIALS_LOGGING_LEVEL >= SUNDIALS_LOGGING_DEBUG
@@ -830,7 +831,7 @@ int lsrkStep_TakeStepRKL(ARKodeMem ark_mem, sunrealtype* dsmPtr, int* nflagPtr)
     }
   }
 
-  step_mem->req_stages = SUNRround(ss);
+  step_mem->req_stages = (int)SUNRround(ss);
   step_mem->stage_max  = SUNMAX(step_mem->req_stages, step_mem->stage_max);
 
 #if SUNDIALS_LOGGING_LEVEL >= SUNDIALS_LOGGING_DEBUG
@@ -1254,7 +1255,7 @@ int lsrkStep_TakeStepSSPs3(ARKodeMem ark_mem, sunrealtype* dsmPtr, int* nflagPtr
   sunrealtype rs  = (sunrealtype)step_mem->req_stages;
   sunrealtype rn  = SUNRsqrt(rs);
   sunrealtype rat = ONE / (rs - rn);
-  int in          = SUNRround(rn);
+  int in          = (int)SUNRround(rn);
 
 #if SUNDIALS_LOGGING_LEVEL >= SUNDIALS_LOGGING_DEBUG
   SUNLogger_QueueMsg(ARK_LOGGER, SUN_LOGLEVEL_DEBUG,
